refactoring library structure
