+++ examples/asciijump/src/main.rs
@@ -2,7 +2,7 @@ use asciijump::game::{SCREEN_HEIGHT, SCREEN_WIDTH};
 use hewn::view::{cursor, ScreenDimensions};
 #[cfg(not(target_arch = "wasm32"))]
 use hewn::{
-    runtime::{initialize_terminal_io, TerminalRuntime},
+    runtime::{modname::initialize_terminal_io, modname::TerminalRuntime},
     view::{TerminalRenderer, View},
 };
 
@@ -16,7 +16,7 @@ fn main() {
 
 #[cfg(not(target_arch = "wasm32"))]
 pub fn play_asciijump_in_terminal() {
-    let (stdout, stdin) = initialize_terminal_io();
+    let (stdout, stdin) = modname::initialize_terminal_io();
     let mut game = create_game(None);
     // TODO where we input height and width as args, can we make it a struct so labelled instead of just
     // guessing?
@@ -32,7 +32,7 @@ pub fn play_asciijump_in_terminal() {
         view_cursor: hewn::view::ViewCoordinate { x: 0, y: 0 },
         cursor_strategy: Box::new(cursor::FollowPlayerYCursorStrategy::new()),
     };
-    let mut runtime = TerminalRuntime::new(stdin, &mut game, &mut display);
+    let mut runtime = modname::TerminalRuntime::new(stdin, &mut game, &mut display);
 
     runtime.start();
 }
