+++ examples/snake/src/lib.rs
@@ -8,6 +8,6 @@ use wasm_bindgen::prelude::*;
 pub fn run_in_canvas(width: u16, height: u16, seed: Option<u64>) {
     let mut game = create_game(width, height, seed);
     game.start_game();
-    let mut runtime = hewn::runtime::WindowRuntime::new();
+    let mut runtime = hewn::runtime::modname::WindowRuntime::new();
     let _ = runtime.start(&mut game);
 }
