+++ src/engine.rs
@@ -1,2 +1,1 @@
 pub mod ecs;
-pub mod view;
