+++ src/terminal.rs
@@ -1,0 +1,4 @@
+#[cfg(not(target_arch = "wasm32"))]
+pub mod render;
+#[cfg(not(target_arch = "wasm32"))]
+pub mod runtime;
